Payment and wallet top up integration

+++ app/Http/Controllers/PricingController.php
@@ -10,6 +10,11 @@ class PricingController extends Controller
 {
     public function index()
     {
+        $admin = auth()->user();
+        if ($admin->role !== 'Administrator'){
+            return back()->with('error', 'Unauthorized access');
+        }
+
         $pricing_categories = PricingCategory::with('pricings')->get();
             
         return view('admin.pricing', compact('pricing_categories'));
@@ -18,6 +23,11 @@ class PricingController extends Controller
     public function store(Request $request)
     {
         try {    
+            $admin = auth()->user();
+            if ($admin->role !== 'Administrator'){
+                return back()->with('error', 'Unauthorized access');
+            }
+            
             $request->validate([
                 'item_name' => 'required|string|unique:pricings',
                 'price' => 'required|string|max:15',
@@ -41,7 +51,12 @@ class PricingController extends Controller
 
     public function update(Request $request, $pricingId)
     {
-        try {    
+        try { 
+            $admin = auth()->user();
+            if ($admin->role !== 'Administrator'){
+                return back()->with('error', 'Unauthorized access');
+            }
+
             $data = $request->validate([
                 'item_name' => 'sometimes|string|unique:pricing',
                 'price' => 'sometimes|string|max:15',
@@ -61,6 +76,11 @@ class PricingController extends Controller
 
     public function destroy($pricingId)
     {
+        $admin = auth()->user();
+        if ($admin->role !== 'Administrator'){
+            return back()->with('error', 'Unauthorized access');
+        }
+        
         $pricing = Pricing::findOrFail($pricingId);
         $pricing->delete();
